Avoid unnecessary reprojection in MVT format
When using ol.render.Feature there was a mandatory geometry transform.
The reprojection now only occurs when an option.featureProjection is defined.

With this change it is now easy again to use the MVT format as a standalone way to read features in the native tile projection.
Since there is no reprojection, there is no need either to define an extent and a world extent, for that use-case.